Address the two remaining review findings that were mine to fix
Retained history is addressable past row 65,535. Ghostty's point
coordinate is a `uint16_t x` and a `uint32_t y`, and its header says the
row "may exceed page size for screen/history tags". The shim declared `y`
as `uint16_t`, so `history()` narrowed every requested row and returned
`BudgetExceeded` for a row that was retained and readable -- the wrong
answer and the wrong error, under a ceiling ADR 0006 never states. The
coordinate is now `u32` through `rt_cell`, the boundary contract's own
prototype, the FFI declaration and both call sites. The active-screen path
widens at the call rather than changing its own bounds, which stay the
grid's.

An incomplete write no longer loses its tail. `WriteOutcome` carries the
accepted byte count and any error, and both were discarded: a ready
operation was treated as a delivered one. If the OS accepted a prefix
before the deadline the rest vanished, which at a prompt looks like the
terminal swallowing what was typed. The submitted chunk is now retained
beside the operation, the unaccepted tail is pushed back to the front of
the queue so the child still sees one ordered stream, and a reported error
fails the session instead of passing silently.

Two findings from the same review are deliberately left open, as
unresolved threads rather than fixes. `Runtime::new` forks to materialize
the helper image and is reached from a worker thread after the
multithreaded Tokio runtime is running; there is no public entry point to
materialize the image before the runtime starts, so closing it needs an
API decision that belongs with the helper-image work, not a restructure
of the demo around a contract I would be guessing at. The demo's style
table also interns every distinct colour for the life of a session, which
a truecolor animation grows without bound; capping it needs a table
generation the client can be told about, and protocol version 1 is frozen.

`python3 scripts/gate.py` passes, including the native rebuild and the C
boundary contract.

+++ crates/infrastructure/src/terminal/ffi.rs
@@ -93,7 +93,11 @@ unsafe extern "C" {
         owner: *mut c_void,
         history: i32,
         x: u16,
-        y: u16,
+        // Ghostty's point coordinate is a u32 row and documents that it "may
+        // exceed page size for screen/history tags". Narrowing it here would
+        // put a 65,535-row ceiling on retained history that nothing else in
+        // the contract states.
+        y: u32,
         text: *mut u32,
         cap: usize,
         len: *mut usize,
